culinary-demo/requirements.txt
llama-index-core>=0.14.12
llama-index-embeddings-google>=0.4.1
llama-index-embeddings-google-genai>=0.3.1
llama-index-vector-stores-pinecone>=0.7.1
google-generativeai>=0.5.4
pinecone>=7.3.0
python-dotenv>=1.2.1
llama-index-llms-cerebras>=0.3.1
streamlit>=1.50.0
nest_asyncio>=1.6.0
tavily-python>=0.7.19
